Adjust index path for cell reuseid in carouselview (#15140)
* Adjust index path for cell reuseid in carouselview

Since CarouselView uses a fake number of items for looping effect like android does, we also need to correct the index used in the base items controller to be the actual item index, not the looped one.

* Fix return

* Remove ifdef for net8, we are net8

This was also masking the public api txt warnings, but have added the unshipped api's as well.